pyproject.toml
[tool.poetry]
name = "test-technique-vif-api-dashboard"
version = "0.1.0"
description = ""
authors = ["zerippeur <zerippeur@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
tensorflow = "^2.11"
tensorflow-io-gcs-filesystem = "^0.26.0"
pandas = "^2.2.2"
tensorflow-hub = "^0.16.1"
numpy = ">=1.23.5,<2.0.0"
mlflow = "^2.14.2"
keras = ">=2.13.1,<2.16"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
streamlit = "^1.36.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
